Fix bugs and add functionality
Fixed problems with the placement of components, in particular, tags going beyond the bottom border of the application, as a result of which the "Request a recipe" and "Exit" buttons were shifted outside the application window.
The ability to exit the application by clicking on the "Exit" button was also added.