Add signal in C++ and slot in QML for data size changed
When filling in data from the data processing class, the model method is
called. This method is used to populate data. After filling in the data,
a signal is emitted about changing the data as well as about changing
the size of the data

The model header file uses `Q_PROPERTY`, which passes a method that
returns the new total amount of data and the signal to be processed in
the QML